Explaination
These discounts form as if it were a "chain", that is, one connected to
the other. Hence the name of the design pattern: Chain of
Responsibility. The idea of the pattern is to solve problems like these:
according to the scenario, we must take some action. Instead of writing
procedural code, and letting a single method find out what to do, we
break those responsibilities into several different classes, and join
them as a chain.